+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Added a test to make sure the derive macros stay compatible with new cw-orch versions
 - Changed the derive macros import from cw_orch to cw_orch_core. This allows changing the cw-orch API without breaking the derive macros.
-
+- Cw-orch mock env info doesn't error when using chain ids that don't match the `osmosis-1` pattern
+- 
 ## 0.22.0
 
 - Updated osmosis test tube to 24.0.1 ,that avoids re-compiling osmosis test tube